pyproject.toml
[project]
name = "aktienticker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.12",
    "ipython>=9.7.0",
    "langchain>=1.1.0",
    "langchain-ollama>=1.0.0",
    "langgraph>=1.0.4",
    "python-dotenv>=1.2.1",
]
